Move short-lived-token files to develocity package
